tests/wpt/web-platform-tests/webxr/xrWebGLLayer_framebuffer.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/webxr_test_constants.js"></script>
<script src="resources/webxr_util.js"></script>
<canvas></canvas>

<script>

let immersiveTestName = "XRWebGLLayer reports a valid framebuffer for immersive sessions";
let inlineTestName = "XRWebGLLayer reports a valid framebuffer for inline sessions";

let fakeDeviceInitParams = TRACKED_IMMERSIVE_DEVICE;

let testFunction = function(session, fakeDeviceController, t) {
  return session.requestReferenceSpace('viewer')
      .then((space) => new Promise((resolve) => {
    function onFrame(time, xrFrame) {
      let layer = xrFrame.session.renderState.baseLayer;
      let gl = layer.context;

      // The layer's framebuffer is a WebGL framebuffer
      t.step(() => {
        assert_not_equals(layer.framebuffer, null);
        assert_true(layer.framebuffer instanceof WebGLFramebuffer);

        // The XR framebuffer is not bound to the GL context by default.
        assert_not_equals(layer.framebuffer, gl.getParameter(gl.FRAMEBUFFER_BINDING));
      });

      // The XR framebuffer can be bound to the GL context.
      gl.bindFramebuffer(gl.FRAMEBUFFER, layer.framebuffer);

      t.step(() => {
        assert_equals(layer.framebuffer, gl.getParameter(gl.FRAMEBUFFER_BINDING));
      });

      // The XR framebuffer has a 2D texture
      let attachment = gl.getFramebufferAttachmentParameter(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.FRAMEBUFFER_ATTACHMENT_OBJECT_TYPE);

      t.step(() => {
        assert_equals(attachment, gl.TEXTURE);
      });

      // Check that each viewport fits inside the framebuffer dimensions
      let viewer_pose = xrFrame.getViewerPose(space);
      for (view of viewer_pose.views) {
        let viewport = layer.getViewport(view);

        t.step(() => {
          assert_less_than_equal(viewport.x + viewport.width, layer.framebufferWidth);
          assert_less_than_equal(viewport.y + viewport.height, layer.framebufferHeight);
        });
      }

      // Finished test.
      resolve();
    }

    session.requestAnimationFrame(onFrame);
  }));
};

xr_session_promise_test(immersiveTestName, testFunction,
  fakeDeviceInitParams, 'immersive-vr');
xr_session_promise_test(inlineTestName, testFunction,
  fakeDeviceInitParams, 'inline');

</script>
