src/third_party/blink/web_tests/external/wpt/webxr/xrWebGLLayer_framebuffer.https-expected.txt

This is a testharness.js-based test.
FAIL XRWebGLLayer reports a valid framebuffer for immersive sessions assert_equals: expected (number) 5890 but got (object) null
FAIL XRWebGLLayer reports a valid framebuffer for inline sessions assert_not_equals: got disallowed value null
Harness: the test ran to completion.
